security student and admin

+++ AdminService/src/main/java/com/adminservice/controller/UserRestController.java
@@ -1,6 +1,7 @@
 package com.adminservice.controller;
 
 import java.net.http.HttpHeaders;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -48,58 +49,82 @@ public class UserRestController {
 	RestTemplate restTemplate;
 
 	public UserRestController() {
+		
 		System.out.println("User Controller constructor called");
 	}
+		//--------------------------------------------------------------------------------Validate-Token-Function----------------------------------------------------------------------------------------------------
 	
-	public boolean validatToken(HttpServletRequest request) {
+		public boolean validateToken(HttpServletRequest request) {
 		
 		org.springframework.http.HttpHeaders headers = new org.springframework.http.HttpHeaders();
 		
 		headers.setContentType(MediaType.APPLICATION_JSON);
 		headers.set("Authorization", "ncs-"+request.getHeader("Authorization"));
-		headers.set("userAuth","admin");
+		headers.set("userType","admin");
+		//headers.set("Authority", request.getHeader("Authority"));
 
 		HttpEntity<String> entity = new HttpEntity<String>(headers);
 		
 		RestTemplate r =new RestTemplate();
-		ResponseEntity<Boolean> tokenTrue=r.exchange("http://localhost:8081/abc-university/public/validate", HttpMethod();
+		ResponseEntity<Boolean> tokenTrue=r.exchange("http://localhost:8081/abc-university/public/validate", HttpMethod.GET,entity,Boolean.class);
 	
 		return tokenTrue.getBody().booleanValue();
-	}
+		}
 		
 		//--------------------------------------------------------------------------------USER----------------------------------------------------------------------------------------------------
 		
 		//Add
 		@PostMapping("/user/add")
-		public ResponseEntity<User> addUser(@RequestBody @Valid User u)
+		public ResponseEntity<User> addUser(@RequestBody @Valid User u,HttpServletRequest request)
 		{
+			boolean tokenValid=validateToken(request);
+			User savedUser =new User();
 			
-			User savedUser = userService.addUser(u);
+			if(tokenValid) {
+				savedUser = userService.addUser(u);
+			}
 			
 			return new ResponseEntity<User>(savedUser,HttpStatus.OK);
 			
 		}
 		//Get All
 		@GetMapping("/user/all")
-		public List<User> getAllUsers()
+		public List<User> getAllUsers(HttpServletRequest request)
 		{
-			return userService.getAllUsers();
+			boolean tokenValid=validateToken(request);
+			List<User> list = new ArrayList<>();
+			
+			if(tokenValid) {
+			list=userService.getAllUsers();
+			}
+			return list;
 		}
 		
 		//Update
 		@PutMapping("/user/update")
-		public ResponseEntity<User> updateProject(@RequestBody User u)
-		{
-			User updateUser = userService.updateUsers(u);
+		public ResponseEntity<User> updateProject(@RequestBody User u, HttpServletRequest request)
+		{	
+			boolean tokenValid=validateToken(request);
+			
+			User updateUser = new User();
+			
+			if(tokenValid) {
+			userService.updateUsers(u);
+			}
+			
 			return new ResponseEntity<>(updateUser,HttpStatus.OK);
 		}
 		
 		
 		//Delete
 		@DeleteMapping("/user/delete/{userId}")
-		public void deleteUser(@PathVariable int userId)
+		public void deleteUser(@PathVariable int userId, HttpServletRequest request)
 		{
-			userService.deleteUserByID(userId);
+			boolean tokenValid=validateToken(request);
+			
+			if(tokenValid) {
+				userService.deleteUserByID(userId);
+			}
 			
 		}
 		
@@ -107,37 +132,58 @@ public class UserRestController {
 		
 		//QUESTION c
 		@PostMapping("/question/add")
-		public ResponseEntity<Question> addQuestion(@RequestBody @Valid Question q)
+		public ResponseEntity<Question> addQuestion(@RequestBody @Valid Question q, HttpServletRequest request)
 		{
+			boolean tokenValid=validateToken(request);
+			Question savedQuestion =new Question();
 			
-			q = restTemplate.postForObject("http://QUESTION-SERVICE/abc-university/result/add", q,Question.class);
-			return new ResponseEntity<Question>(q,HttpStatus.OK);
+			if(tokenValid) {
+				savedQuestion = restTemplate.postForObject("http://QUESTION-SERVICE/abc-university/question/add", q,Question.class);
+			}
+			return new ResponseEntity<Question>(savedQuestion,HttpStatus.OK);
 			
 		}
 		//QUESTION r
 		@GetMapping("/question/all")
-		public List<Question> getAllQuestions()
-		{
-			List<Question> q = restTemplate.getForObject("http://QUESTION-SERVICE/abc-university/question/all/",List.class);
+		public List<Question> getAllQuestions(HttpServletRequest request)
+		{	
+			boolean tokenValid=validateToken(request);
+			
+			List<Question> q = new ArrayList<>();
+			
+			if(tokenValid) {
+				q = restTemplate.getForObject("http://QUESTION-SERVICE/abc-university/question/all",List.class);
+			}
 			return q;
 		}
 		
 		//QUESTION u
 		@PutMapping("/question/update")
-		public ResponseEntity<Question> updateQuestion(@RequestBody Question q)
+		public ResponseEntity<Question> updateQuestion(@RequestBody Question q,HttpServletRequest request)
 		{
-			Question updatedQuestion = q;
-			restTemplate.put("http://QUESTION-SERVICE/abc-university/result/update", updatedQuestion, Question.class);
+			
+			boolean tokenValid=validateToken(request);
+			
+			Question updatedQuestion = new Question();
+			
+			if(tokenValid) {
+				
+				restTemplate.put("http://QUESTION-SERVICE/abc-university/result/update", q, Question.class);
+				updatedQuestion=q;
+			}
 			return new ResponseEntity<Question>(updatedQuestion,HttpStatus.OK);
 		}
 		
 		
 		//QUESTION d
 		@DeleteMapping("/question/delete/{questionId}")
-		public void deleteQuestion(@PathVariable int questionId)
+		public void deleteQuestion(@PathVariable int questionId,HttpServletRequest request)
 		{
-			restTemplate.delete("http://QUESTION-SERVICE/abc-university/question/delete/"+questionId,Question.class);
+			boolean tokenValid=validateToken(request);
 			
+			if(tokenValid) {
+			restTemplate.delete("http://QUESTION-SERVICE/abc-university/question/delete/"+questionId,Question.class);
+			}
 			
 		}
 		
@@ -145,26 +191,47 @@ public class UserRestController {
 		
 		//RESULT c
 		@PostMapping("/result/add")
-		public Result addResult(@RequestBody @Valid Result r) {
-			Result q = restTemplate.postForObject("http://RESULT-SERVICE/abc-university/result/add", r,Result.class);
-			return q;
+		public Result addResult(@RequestBody @Valid Result r,HttpServletRequest request) {
+			boolean tokenValid=validateToken(request);
+			
+			Result savedResult =new Result();
+			
+			if(tokenValid) {
+				savedResult = restTemplate.postForObject("http://RESULT-SERVICE/abc-university/result/add", r,Result.class);
+				
+			}
+			return savedResult;
 		}
 		
 		//RESULT r
 		@GetMapping("/result/all")
-		public List<Result> getAllResult(){
+		public List<Result> getAllResult(HttpServletRequest request){
 			
-			List<Result> r = restTemplate.getForObject("http://RESULT-SERVICE/abc-university/result/all", List.class);
+			boolean tokenValid=validateToken(request);
+			
+			List<Result> r = new ArrayList<>();
+			
+			if(tokenValid) {
+				r = restTemplate.getForObject("http://RESULT-SERVICE/abc-university/result/all", List.class);
+			}
 			return r;
 		}
 		
 		//RESULT u
 		@PutMapping("/result/update")
-		public Result updateResult(@RequestBody Result r) {
+		public Result updateResult(@RequestBody Result r,HttpServletRequest request) {
 			
-			Result updatedResult = r;
-			restTemplate.put("http://RESULT-SERVICE/abc-university/result/update", updatedResult, Result.class);
-			return(r);
+			boolean tokenValid=validateToken(request);
+			
+			Result updatedResult = new Result();
+			
+			if(tokenValid) {
+				
+				restTemplate.put("http://RESULT-SERVICE/abc-university/result/update", updatedResult, Result.class);
+				updatedResult = r;
+			}
+			
+			return(updatedResult);
 		}
